taintfox: Fixing console message issues
Removing browser message taint reporting which was only
reporting DOM sinks, and not for example, from eval().
This closes #4 as the message is no longer reported.

Adding a JS_ReportWarningUTF8() message to JS_ReportTaintSink()
which will report all taint sinks (including eval()), and
hence fixes #7.

As the JS_ReportWarningUTF8() function includes the JavaScript
context, the location of the tainted sink is included,
which fixes #3.