taintfox: Adding localStorage taint source
Adding new localStorage as taint source for client-side stored XSS
detection. This involves addition of sources in three implementations:

1. LocalStorage.cpp - this is the default local storage implementation
2. LSObject.cpp - this is marked as the next-gen implementation
3. ParitionedLocalStorage.cpp - private browsing local storage

In addition, the cache mechanisms for 1. and 2. need modification to
check if the string is tainted, and to update the cache if a taint is
detected. Otherwise replacement of a string with the same string with
taint will not be performed, and the taint will be lost.

In all cases a string retreived with window.localStorage.getItem("key")
will be marked as a taint source if the stored item is not already tainted.
If it is already tainted, the taint flow will be augmented with the "localStorage"
TaintOperation.

This commit fixes #15