taintfox: Changing JSString taint location
The GC expects the first word of every Cell (from which
JSString inherits) to be reserved.

JSString was inheriting from TaintableString, which meant that
the taint_ member was at the start of the struct. This was causing
some problems with the GC, as String type checks were failing when
they should have succeeded.

JSString now contains a StringTaint member after the length parameter
so the reserved bits are now at the start of the memory as expected.

Also added some helper methods to JSString to maintain the interface,
and updated the README with some debugging tips.