taintfox: adding location to taint flows
Adding a TaintLocation class to record the file, function, line
and column of a TaintOperation or TaintSource. This is deliberately
kept free of any SpiderMonkey or XPC classes.

All TaintOperations have been updated to include the TaintLocation
extracted from the JavaScript context. This fixes #29.

The same has been done on the DOM side, to add location information
to all TaintSources.

The taint source and sink reporting has been enhanced to accept an
argument to add extra information about the source or sink. In
particular, the innerHTML sink and input.value sources now report
the "id" of the element in their arguments. This fixes #31.

Additionally fixed style mismatches and added some wrapper functions
to clean up the code.