Performance Tweaks
After a bit of profiling I noticed that creating TaintOperations is very
expensive. In hindsight this is obvious, as we have to walk the stack
to get the location etc.
We did however create a lot of TaintOperation objects without ever using
them, i.e., due to the string itself being untainted.

This change does add explicit checks where possible (in some cases we
create the TaintOperation object in advance due to GC issues, this is
not easily hidden behind a check) so TaintOperations are only created if
the string is actually tainted, i.e., we use the TaintOperation to
extend the TaintFlow.