Unbroke two tests
These tests used to be broken in Foxhound and were fixed in #151.

This fix was intended to mask the fact that these tests were picking up
and failing due to foxhound induced side effects on untainted strings.

As we call toString() on objects during TaintOperation creation, the
usage of Foxhound can have visible side effects. This is shown here as
these tests count how often toString() is called on an object.

Now, in this branch we do not create any TaintOperation objects for
these specific cases of untainted strings anymore. Consequently, the
tests are now passing in their original form.