Foxhound: Fixing GC crashes due to JSONParser strings
The CurrentJsonPath method was causing tab crashes when
parsing large JSON files. These seem to occur when the JSON
parser was running at the same time as Nursery Strings were
moved to Tenured space. My feeling is that the JSONPath strings
were being created faster than they could be moved out of the
nursery, causing some nasty segfaults.

A quick for this appears to be to create the JSONPath strings
directly in the Tenured heap. This is probably not ideal, as
these strings are only used to create TaintOperations.

This is probably worth revisiting to see if there is a better
solution. There seems to be a lot of upcoming upstream changes
related to String GC, so probably worth waiting for the next
release (see e.g. 7e2b43ba2a92968f752a60f90448ea47b75bf66e)