Add build script for project foxhound.
Currently one of the difficulties for people to get started with project
foxhound seems to be the build. As how to build foxhound with playwright
is something that's currently undocumented outside of the SAP CI and my
adaption of it, this seems like a big hurdle to overcome.

This PR aims to be the first step to automate most of the build
requirements. It borrows heavily from Foxhound_builder, but as this
script is supposed to live in tree we can simplify stuff greatly.

We can, for example, make the playwright version available in machine
readable format. This avoids the issue that each user has to update the build
script on each release, because we can maintain the compatible
playwright version. This can be extended to cover checks for usable
rustc versions in the future.

Generally, there are some additions that would be nice to have in the
future:

- Support to make a commit for the version we are trying to build.
This matters in combination with playwright, as we avoid having an unclean
tree. It also brings some issues, as we would have to detect that we are
building on top of a branch used to build playwright to avoid throwing
the branch away for each build (?)

- Support to disable the preparation phase.
If we want to test small changes iteratively, re-applying the
playwright patches, running bootstrap etc is superfluous and can be
avoided.

- Additional safeguards.
Such as tests for the correct rust version. This might be cumbersome as
we would have to ensure the script picks up the same rust version mach
does. It seems to favor the rustup based version over the version that
is installed on the main OS. This requires testing, as this is only
based on my observations so far.

- Post build checks.

For example, if we compile with playwright support, we can test
`foxhound -juggler-pipe` and if the flag is recognized, the playwright
integration is available.