Make StructuredClone Taint Aware
This change adds taint support to the StructuredClone machinery.

Before, when directly invoking `window.structuredClone()` to copy an object or indirectly, by sending a tainted value via postMessage, the taint was lost.

This adds support to StructuredClone to send tainted values. It works as follows:

Based on the TaintSpew code I extracted a generic method to serialize `StringTaint` objects as JSON. When a tainted String is getting cloned, we detect this and instead send a Tainted String, which essentially attaches a second string, containing the serialized taint object.
On the receiver side, we detect this, deserialize the JSON to a `StringTaint` (the majority of this PR actually) and attach it to the returned string.

IN JS shell this can be tested as follows:

```javascript
o = serialize(String.tainted("foo") + " - " + String.tainted("bar"));
o2 = deserialize(o);
o2.taint
o3 = serialize({k: String.tainted("bar"), kkk: "value"});
o4 = deserialize(o3);
o4.k.taint
```