Allowing to apply patches with build.sh
Currently the way the build script works, especially when compiling with Playwright support is that it tries to ensure that the Git worktree is clean, i.e., has no uncommitted changes. This aims to prevent that by accident one ends up in a situation where substantial changes have been made and the build script happily applies the Playwright patches. Reverting them requires a bit of understanding of the underlying build, so we try to safeguard against this.

However, this is at times a bit annoying. One could always set an environment variable to skip the check, but there are situation where that is cumbersome. Consider for example, you'd want to remove the `ChromeOnly` attribute from the [DOMParser.parseFromSafeString()](https://github.com/SAP/project-foxhound/blob/main/dom/webidl/DOMParser.webidl#L30). This is an easy change, but if your memory is as bad as mine remembering in which folder to locate the change when redoing it after an update is annoying. This can be worked around with Git stash trickery, but I feel like we can offer a nicer way to do this. This PR aims to implement the nicer approach.

The core idea is, that you can provide the build script with a sequence of patch files, which are applied before and reverted after a successful build. This ensures your working tree is clean, and allows you to apply changes rather easily.

The workflow now becomes as follows:
1. Remove the troublesome attribute and save the webidl file.
2. `git diff >> ../patches/expose_safe_parse_to_webpages.patch`
3. `FOXHOUND_PATCHES="$(pwd)/../expose_safe_parse_to_webpages.patch" bash build.sh -ap`

Several patches can be passed to `$FOXHOUND_PATCHES` separated by semicolons.

The script tries to ensure that all provided patches exist and errors out if that is not the case or they fail to apply, so it aims to make this as bulletproof as possible.