Workaround for nsPipe Change
As we can not get a direct reference to the pipe anymore, this commit adds an workaround.
We can now mark inputstreams as taintable, i.e., they support the `SetTaint method`.
As the input stream can reach back into the pipe, it can reuse the old taint propagation machinery as before.

This lacks meaningful testing, as mochitests can't capture all that well that a piece of HTML from the server is tainted.
When enabling nsHTTP logging, we can see the taint getting passed along however.